Move some Dataset CRUD related code to separate methods
